주문오류 수정 / 주문취소 기능 구현

+++ src/main/java/spring/controller/OrderController.java
@@ -1,5 +1,6 @@
 package spring.controller;
 
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -10,10 +11,15 @@ import javax.servlet.http.HttpSession;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
+
+import com.siot.IamportRestClient.IamportClient;
+import com.siot.IamportRestClient.exception.IamportResponseException;
+import com.siot.IamportRestClient.request.CancelData;
 
 import spring.dao.CategoryDao;
 import spring.dao.OrderDao;
@@ -35,6 +41,12 @@ public class OrderController {
 
 	public void setDao(OrderDao dao) {
 		this.dao = dao;
+	}
+	
+	private IamportClient api;
+	
+	public OrderController() {
+		this.api = new IamportClient("5478353111638089","38c701ccf0c5e1bb14f091d942224863eebfa6b285a8195735b0eaae973d6339abf549d563d49cf8");
 	}
 
 	@RequestMapping(value = "/payment", method = RequestMethod.POST)
@@ -134,6 +146,7 @@ public class OrderController {
 		
 		return "order/order";
 	}
+	
 	@RequestMapping(value = "/cancel", method = RequestMethod.GET)
 	public String cancelG(Model model,HttpSession session, HttpServletRequest request) {
 		
@@ -145,12 +158,56 @@ public class OrderController {
 		return "redirect:/product/cart/list.do";
 	}
 	
-	@RequestMapping(value = "/orderInsert", method = RequestMethod.POST)
-	public String OrderInsert32P(Model model, Order order,HttpSession session, HttpServletRequest request) {
+	@RequestMapping(value = "/paymentCancle/{order_number}", method = RequestMethod.GET)
+	public String paymentCancelG(@PathVariable("order_number") int orderNum,Model model,HttpSession session, HttpServletRequest request) throws IamportResponseException, IOException{
 		
+		Order order = dao.orderinfo(orderNum);
+
+		if(order.getPay_status().equals("결제완료")) {
+			CancelData cancel = new CancelData(order.getImp_uid(),true);
+			api.cancelPaymentByImpUid(cancel);
+		}else {
+			model.addAttribute("msg", "결제가 취소되어 있는 상품입니다.");
+			return "redirect:/order/orderStatus";
+		}
+		
+		order.setOrder_sub((ArrayList<OrderSub>) dao.productListinfo(order.getOrder_join_number()));
+		ArrayList<Option> optionData = new ArrayList<Option>();
+		for(int i=0; i<order.getOrder_sub().size();i++) {
+			optionData = (ArrayList<Option>) dao.optionListinfo(order.getOrder_sub().get(i).getOption_join_number());
+			order.getOrder_sub().get(i).setOption_sub(optionData);
+		}
+	
+		//수량 원복
+		for(int i=0; i < order.getOrder_sub().size();i++) {
+			dao.productRollBack(order.getOrder_sub().get(i));
+			for(Option option : order.getOrder_sub().get(i).getOption_sub()) {
+				if(option.getOption_Count() !=0) {
+					dao.optionRollBack(option);
+				}
+			}
+		}
+		//포인트 원복
+		if(order.getUse_point()!=0) {
+			dao.pointRollBack(order);
+		}
+		
+		//결제취소로 변경
+		dao.payment_status_edit(orderNum);
+		
+		return "redirect:/order/orderStatus";
+	}
+	
+	@RequestMapping(value = "/orderInsert", method = RequestMethod.POST)
+	public String OrderInsertP(Model model, Order order,HttpSession session, HttpServletRequest request) {
+		
+
 		for(OrderSub deleteData : order.getOrder_sub()) {
 			dao.deleteCart(deleteData.getCartoption_number());
-			dao.deleteCartOption(deleteData.getCartoption_number());
+			if(deleteData.getOption_sub()!=null) {
+				dao.deleteCartOption(deleteData.getCartoption_number());	
+			}
+			
 		}
 		dao.insertOrder(order);
 		//joinnum
@@ -163,16 +220,16 @@ public class OrderController {
 			dao.insertProductList(order.getOrder_sub().get(i));
 			dao.productDeduction(order.getOrder_sub().get(i));
 			option_join_number = dao.option_join_number(order_join_number);
-			for(Option option : order.getOrder_sub().get(i).getOption_sub()) {
-				System.out.println(option.getOption_Count());
-				System.out.println("여기");
-				if(option.getPayment_option_count() !=0) {
-					option.setOption_Join_Number(option_join_number);
-					dao.insertOptionList(option);
-					dao.optionDeduction(option);
+			if(order.getOrder_sub().get(i).getOption_sub()!=null) {
+				for(Option option : order.getOrder_sub().get(i).getOption_sub()) {
+					if(option.getPayment_option_count() !=0) {
+						option.setOption_Join_Number(option_join_number);
+						dao.insertOptionList(option);
+						dao.optionDeduction(option);
+					}
 				}
-				
 			}
+			
 		}
 		//포인트 차감
 		if(order.getUse_point()!=0) {
@@ -207,7 +264,9 @@ public class OrderController {
 			info.get(i).setOrder_sub(productData);
 			for(int j=0; j<info.get(i).getOrder_sub().size();j++) {
 				optionData = (ArrayList<Option>) dao.optionListinfo(info.get(i).getOrder_sub().get(j).getOption_join_number());
-				info.get(i).getOrder_sub().get(j).setOption_sub(optionData);
+				if(info.get(i).getOrder_sub().get(j).getOption_sub()!=null) {
+					info.get(i).getOrder_sub().get(j).setOption_sub(optionData);
+				}
 			}
 		}
